moved debug commands to separate module
